解决 mingw64 x64 编译错误问题

+++ utils.h
@@ -10,7 +10,7 @@
 #define ARRAYSIZE(a) (sizeof((a)) / sizeof((a)[0]))
 #endif
 
-#define offset_of(type, member) ((int)&((type*)0)->member)
+#define offset_of(type, member) ((size_t)&((type*)0)->member)
 #define container_of(ptr, type, member) (type*)((char*)(ptr) - offset_of(type, member))
 
 uint32_t get_tick_count(void);
